# backend/internal/dashboard/data/openapi.json
{
  "components": {
    "schemas": {
...
      "RestartResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "message",
          "ok"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "fb_admin",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Dashboard admin surface. Generated by backend/cmd/openapi-emit from admin_manifest.json + dashboard.AdminAPIPaths — do not edit by hand.",
    "title": "freebucks-proxy admin API",
    "version": "dev"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/admin/restart": {
      "post": {
        "operationId": "restart",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestartResponse"
                }
              }
            },
            "description": "Validate config and restart the gateway process"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Validate config and restart the gateway process",
        "tags": [
          "system"
        ]
      }
    },
...
  }
}